jc-rpc-common/src/test/java/com/jcshang/jcrpc/common/utils/ReflectionUtilsTest.java
package com.jcshang.jcrpc.common.utils;

import junit.framework.TestCase;
import org.junit.Test;

import java.lang.reflect.Method;

import static org.junit.Assert.*;

public class ReflectionUtilsTest {

    @Test
    public void newInstance() {
        TestClass t = ReflectionUtils.newInstance(TestClass.class);
        assertNotNull(t);
    }

    @Test
    public void getPublicMethods() {
        Method[] methods = ReflectionUtils.getPublicMethods(TestClass.class);
        assertEquals(1, methods.length);

        String methodName = methods[0].getName();
        assertEquals("b", methodName);
    }

    @Test
    public void invoke() {
        Method b = ReflectionUtils.getPublicMethods(TestClass.class)[0];
        TestClass t = new TestClass();
        Object res = ReflectionUtils.invoke(t, b);
        assertEquals("b", res);
    }
}
